Working on drawing menus
I broke out menu parts into their own files so i could find things easier.

fixed a text rendering bug from converting byte to string. we need byte values to live to the rendering stage.
https://github.com/MaKiPL/OpenVIII/issues/54
