+++ FF8/module_main_menu_debug.cs
@@ -1,42 +1,29 @@
 ﻿using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Text;
 
 namespace FF8
 {
-    internal static class Module_main_menu_debug
+    internal static partial class Module_main_menu_debug
     {
         #region Fields
 
-        /// <summary>
-        /// Strings for the debug menu
-        /// </summary>
-        private static Dictionary<Enum, Item> strDebugLobby;
-
-        /// <summary>
-        /// Strings for the main menu
-        /// </summary>
-        private static Dictionary<Enum, Item> strMainLobby;
-
-        private static int debug_choosedBS, debug_choosedAudio;
-        private static string debug_choosedField = Memory.FieldHolder.fields[debug_fieldPointer];
-        private static string debug_choosedMovie = Path.GetFileNameWithoutExtension(Module_movie_test.Movies[debug_moviePointer]);
-        private static string debug_choosedMusic = Path.GetFileNameWithoutExtension(Memory.dicMusic[0][0]);
-        private static int debug_fieldPointer = 90;
-        private static int debug_moviePointer = 0;
         private static float fade;
+
+        //        case 1:
+        //            filename = aw.GetListOfFiles().First(x => x.ToLower().Contains("start01"));
+        //            break;
+        //    }
+        //    tex = new TEX(ArchiveWorker.GetBinaryFile(Memory.Archives.A_MENU, filename));
+        //    return TextureHandler.UseBest(tex,TextureHandler.LoadPNG(filename));
+        //}
+        private static bool LastActive = false;
 
         //private static float fScaleHeight;
         //private static float fScaleWidth;
         private static float lastfade;
 
-        private static Ditems s_dchoose;
-        private static Mitems s_mchoose;
-        private static TextureHandler start;
         private static MainMenuStates State = 0;
         private static int vpHeight;
         private static int vpWidth;
@@ -45,22 +32,14 @@ namespace FF8
 
         #region Enums
 
-        /// <summary>
-        /// Identifiers and Ordering of debug menu items
-        /// </summary>
-        private enum Ditems
+        private enum Litems
         {
-            Reset,
-            Overture,
-            Battle,
-            Field,
-            Movie,
-            Music,
-            Sounds,
-            World,
-            Faces,
-            Icons,
-            Cards
+            GameFolder,
+            Load,
+            LoadFF8,
+            Loading,
+            Slot1,
+            Slot2
         }
 
         /// <summary>
@@ -72,91 +51,17 @@ namespace FF8
             MainLobby,
             NewGameChoosed,
             LoadGameLoading,
-            LoadGameScreen,
+            LoadGameChooseSlot,
+            LoadGameChooseGame,
             DebugScreen
-        }
-
-        /// <summary>
-        /// Identifiers and Ordering of main menu items
-        /// </summary>
-        private enum Mitems
-        {
-            New,
-            Load,
-            Debug
         }
 
         #endregion Enums
 
         #region Properties
 
+        public static float vpSpace { get; private set; }
         private static float Fade { get => fade; set => fade = value; }
-        private static Vector2 Offset { get; set; }
-
-        /// <summary>
-        /// Currently selected Field
-        /// </summary>
-        public static int FieldPointer
-        {
-            get => debug_fieldPointer;
-            set
-            {
-                if (value >= Memory.FieldHolder.fields.Length)
-                {
-                    value = 0;
-                }
-                else if (value < 0)
-                {
-                    value = Memory.FieldHolder.fields.Length - 1;
-                }
-
-                debug_fieldPointer = value;
-                Memory.FieldHolder.FieldID = (ushort)value;
-                debug_choosedField = Memory.FieldHolder.fields[value];
-            }
-        }
-
-        /// <summary>
-        /// Currently selected Movie
-        /// </summary>
-        public static int MoviePointer
-        {
-            get => debug_moviePointer;
-            set
-            {
-                if (value >= Module_movie_test.Movies.Count)
-                {
-                    value = 0;
-                }
-                else if (value < 0)
-                {
-                    value = Module_movie_test.Movies.Count - 1;
-                }
-
-                debug_moviePointer = value;
-                Module_movie_test.Index = value;
-                debug_choosedMovie = Path.GetFileNameWithoutExtension(Module_movie_test.Movies[value]);
-            }
-        }
-
-        /// <summary>
-        /// Current choice on debug menu
-        /// </summary>
-        private static Ditems Dchoose
-        {
-            get => s_dchoose; set
-            {
-                if (value > s_dchoose && value > (Ditems)Enum.GetValues(typeof(Ditems)).Cast<int>().Max())
-                {
-                    value = 0;
-                }
-                else if (value < s_dchoose && s_dchoose <= 0)
-                {
-                    value = (Ditems)Enum.GetValues(typeof(Ditems)).Cast<int>().Max();
-                }
-                s_dchoose = value;
-            }
-        }
 
         /// <summary>
         /// Current choice on main menu
@@ -178,69 +83,53 @@ namespace FF8
             }
         }
 
-        public static float vpSpace { get; private set; }
-        public static Vector2 DFontPos { get; private set; }
+        private static Dictionary<Enum, Item> strLoadScreen { get; set; }
 
         #endregion Properties
 
         #region Methods
 
         /// <summary>
-        /// Init
+        /// Trigger required draw function.
         /// </summary>
-        private static void Init()
+        internal static void Draw()
         {
-            strDebugLobby = new Dictionary<Enum, Item>()
+            Memory.graphics.GraphicsDevice.Clear(Color.Black);
+            //fScaleWidth = (float)Memory.graphics.GraphicsDevice.Viewport.Width / Memory.PreferredViewportWidth;
+            //fScaleHeight = (float)Memory.graphics.GraphicsDevice.Viewport.Height / Memory.PreferredViewportHeight;
+            vpWidth = Memory.PreferredViewportWidth;//Memory.graphics.GraphicsDevice.Viewport.Width;
+            vpHeight = Memory.PreferredViewportHeight;//Memory.graphics.GraphicsDevice.Viewport.Width;
+            lastfade = fade;
+            vpSpace = vpHeight * 0.09f * Memory.Scale().X;
+            DFontPos = new Vector2(vpWidth * .10f * Memory.Scale().X, vpHeight * .05f * Memory.Scale().Y) + Offset;
+            switch (State)
             {
-                { Ditems.Reset, new Item{Text=Font.CipherDirty("Reset Main Menu state") } },
-                { Ditems.Overture, new Item{Text=Font.CipherDirty("Play Overture")} },
-                { Ditems.Battle, new Item{Text=Font.CipherDirty("Battle encounter: ")+"{0}"} },
-                { Ditems.Field, new Item{Text=Font.CipherDirty("Field debug render: ")+"{0}"} },
-                { Ditems.Movie, new Item{Text=Font.CipherDirty("Movie debug render: ")+"{0}"} },
-                { Ditems.Music, new Item{Text=Font.CipherDirty("Play/Stop music: ")+"{0}"} },
-                { Ditems.Sounds, new Item{Text=Font.CipherDirty("Play audio.dat: ")+"{0}"} },
-                { Ditems.World, new Item{Text=Font.CipherDirty("Jump to World Map")} },
-                { Ditems.Faces, new Item{Text=Font.CipherDirty("Test Faces")} },
-                { Ditems.Icons, new Item{Text=Font.CipherDirty("Test Icons")} },
-                { Ditems.Cards, new Item{Text=Font.CipherDirty("Test Cards")} }
-            };
-            strMainLobby = new Dictionary<Enum, Item>()
-            {
-                { Mitems.New, new Item{Text=Encoding.UTF8.GetString(Memory.Strings.Read(Strings.FileID.MNGRP, 1 ,105)) } },
-                { Mitems.Load, new Item{Text=Encoding.UTF8.GetString(Memory.Strings.Read(Strings.FileID.MNGRP, 1 ,106)) } },
-                { Mitems.Debug, new Item{Text=Font.CipherDirty("OpenVIII debug tools") } }
-            };
-            Memory.Strings.Close();
-            if (start == null)
-            {
-                start = new TextureHandler("start{0:00}", 2);
+                case MainMenuStates.Init:
+                case MainMenuStates.MainLobby:
+                    DrawMainLobby();
+                    break;
+
+                case MainMenuStates.DebugScreen:
+                    DrawDebugLobby();
+                    break;
+
+                case MainMenuStates.NewGameChoosed:
+                    DrawMainLobby();
+                    break;
+
+                case MainMenuStates.LoadGameChooseSlot:
+                    DrawLGChooseSlot();
+                    break;
+
+                case MainMenuStates.LoadGameChooseGame:
+                    DrawLGChooseGame();
+                    break;
+
+                case MainMenuStates.LoadGameLoading:
+                    DrawLoading();
+                    break;
             }
         }
-
-        ///// <summary>
-        ///// Get reqeusted texture from Menu archive.
-        ///// </summary>
-        ///// <param name="v">0 or 1</param>
-        ///// <returns>Requested Texture</returns>
-        //private static Texture2D GetTexture(int v)
-        //{
-        //    ArchiveWorker aw = new ArchiveWorker(Memory.Archives.A_MENU);
-        //    TEX tex;
-        //    string filename = "";
-        //    switch (v)
-        //    {
-        //        case 0:
-        //            filename = aw.GetListOfFiles().First(x => x.ToLower().Contains("start00"));
-        //            break;
-
-        //        case 1:
-        //            filename = aw.GetListOfFiles().First(x => x.ToLower().Contains("start01"));
-        //            break;
-        //    }
-        //    tex = new TEX(ArchiveWorker.GetBinaryFile(Memory.Archives.A_MENU, filename));
-        //    return TextureHandler.UseBest(tex,TextureHandler.LoadPNG(filename));
-        //}
-        private static bool LastActive = false;
 
         /// <summary>
         /// Triggers functions depending on state
@@ -293,430 +182,24 @@ namespace FF8
                     UpdateNewGame();
                     break;
 
-                case MainMenuStates.LoadGameLoading:
+                case MainMenuStates.LoadGameChooseSlot:
+                    UpdateLGChooseSlot();
+                    Memory.IsMouseVisible = true;
                     break;
 
-                case MainMenuStates.LoadGameScreen:
+                case MainMenuStates.LoadGameChooseGame:
+                    UpdateLGChooseGame();
                     Memory.IsMouseVisible = true;
+                    break;
+
+                case MainMenuStates.LoadGameLoading:
+                    UpdateLoading();
+                    Memory.IsMouseVisible = false;
                     break;
 
                 default:
                     goto case 0;
             }
-        }
-
-        /// <summary>
-        /// Update Main Lobby
-        /// </summary>
-        /// <returns>true on change</returns>
-        private static bool UpdateMainLobby()
-        {
-            bool ret = false;
-            Point ml = Input.MouseLocation;
-            foreach (KeyValuePair<Enum, Item> entry in strMainLobby)
-            {
-                if (entry.Value.Loc.Contains(ml))
-                {
-                    Mchoose = (Mitems)entry.Key;
-                    ret = true;
-
-                    if (Input.Button(Buttons.MouseWheelup) || Input.Button(Buttons.MouseWheeldown))
-                    {
-                        return ret;
-                    }
-                    break;
-                }
-            }
-            if (Input.Button(Buttons.Down))
-            {
-                Input.ResetInputLimit();
-                init_debugger_Audio.PlaySound(0);
-                Mchoose++;
-                ret = true;
-            }
-            if (Input.Button(Buttons.Up))
-            {
-                Input.ResetInputLimit();
-                init_debugger_Audio.PlaySound(0);
-                Mchoose--;
-                ret = true;
-            }
-            if (Input.Button(Buttons.Okay))
-            {
-                Input.ResetInputLimit();
-                switch (Mchoose)
-                {
-                    case Mitems.New:
-                        init_debugger_Audio.PlaySound(28);
-                        State = MainMenuStates.NewGameChoosed;
-                        break;
-                    //case Mitems.Load:
-                    //    init_debugger_Audio.PlaySound(0);
-                    //    State = MainMenuStates.LoadGameScreen;
-                    //    break;
-                    case Mitems.Debug:
-                        init_debugger_Audio.PlaySound(0);
-                        Mchoose = 0;
-                        Dchoose = 0;
-                        State = MainMenuStates.DebugScreen;
-                        Fade = 0.0f;
-                        break;
-                }
-                ret = true;
-            }
-            return ret;
-        }
-
-        /// <summary>
-        /// Waits for fade to end then triggers new game.
-        /// </summary>
-        private static void UpdateNewGame()
-        {
-            if (Fade > 0.0f)
-            {
-                Fade -= Memory.gameTime.ElapsedGameTime.Milliseconds / 1000.0f / 2;
-                return;
-            }
-            /*reverse engineering notes:
-            *
-            * we should happen to reset wm2field values
-            * also the basic party of Squall is now set: SG_PARTY_FIELD1 = 0, and other members are 0xFF
-            */
-            FieldPointer = 74; //RE: startup stage ID is hardcoded. Probably we would want to change it for modding
-            //the module changes to 1 now
-            Module_field_debug.ResetField();
-
-            Module_movie_test.Index = 30;
-            Module_movie_test.ReturnState = Memory.MODULE_FIELD_DEBUG;
-            Memory.module = Memory.MODULE_MOVIETEST;
-            State = MainMenuStates.MainLobby;
-        }
-
-        /// <summary>
-        /// Update Debug Menu
-        /// </summary>
-        /// <returns>true on change</returns>
-        private static bool UpdateDebugLobby()
-        {
-            bool ret = false;
-            Point ml = Input.MouseLocation;
-            foreach (KeyValuePair<Enum, Item> entry in strDebugLobby)
-            {
-                if (entry.Value.Loc.Contains(ml))
-                {
-                    Dchoose = (Ditems)entry.Key;
-                    ret = true;
-
-                    if (Input.Button(Buttons.MouseWheelup))
-                    {
-                        return UpdateDebugLobbyLEFT();
-                    }
-                    if (Input.Button(Buttons.MouseWheeldown))
-                    {
-                        return UpdateDebugLobbyRIGHT();
-                    }
-                    break;
-                }
-            }
-            if (Input.Button(Buttons.Down))
-            {
-                Input.ResetInputLimit();
-                init_debugger_Audio.PlaySound(0);
-                Dchoose++;
-                ret = true;
-            }
-            if (Input.Button(Buttons.Up))
-            {
-                Input.ResetInputLimit();
-                init_debugger_Audio.PlaySound(0);
-                Dchoose--;
-                ret = true;
-            }
-            if (Input.Button(Buttons.Okay) && Dchoose == Ditems.Reset || Input.Button(Buttons.Cancel))
-            {
-                Input.ResetInputLimit();
-                init_debugger_Audio.PlaySound(8);
-                init_debugger_Audio.StopAudio();
-                Dchoose = 0;
-                Fade = 0.0f;
-                State = MainMenuStates.MainLobby;
-                ret = true;
-            }
-            else if (Input.Button(Buttons.Okay))
-            {
-                ret = UpdateDebugLobbyOKAY();
-            }
-            else if (Input.Button(Buttons.Left))
-            {
-                ret = UpdateDebugLobbyLEFT();
-            }
-            else if (Input.Button(Buttons.Right))
-            {
-                ret = UpdateDebugLobbyRIGHT();
-            }
-            return ret;
-        }
-
-        private static bool UpdateDebugLobbyLEFT()
-        {
-            bool ret = true;
-            Input.ResetInputLimit();
-            switch (Dchoose)
-            {
-                case Ditems.Sounds:
-                    if (debug_choosedAudio > 0)
-                    {
-                        debug_choosedAudio--;
-                    }
-
-                    break;
-
-                case Ditems.Battle:
-                    if (debug_choosedBS <= 0)
-                    {
-                        return false;
-                    }
-
-                    debug_choosedBS--;
-                    break;
-
-                case Ditems.Field:
-                    FieldPointer--;
-                    break;
-
-                case Ditems.Movie:
-                    MoviePointer--;
-                    break;
-
-                case Ditems.Music:
-
-                    if (Memory.MusicIndex <= ushort.MinValue)
-                    {
-                        Memory.MusicIndex = ushort.MaxValue;
-                    }
-                    else
-                    {
-                        Memory.MusicIndex--;
-                    }
-                    debug_choosedMusic = Path.GetFileNameWithoutExtension(Memory.dicMusic[Memory.MusicIndex][0]);
-                    break;
-
-                default:
-                    ret = false;
-                    break;
-            }
-            if (ret)
-            {
-                init_debugger_Audio.PlaySound(0);
-            }
-
-            return ret;
-        }
-
-        private static bool UpdateDebugLobbyOKAY()
-        {
-            bool ret = true;
-            Input.ResetInputLimit();
-            switch (Dchoose)
-            {
-                case Ditems.Overture:
-                    Dchoose = 0;
-                    Fade = 0.0f;
-                    State = MainMenuStates.MainLobby;
-                    Module_overture_debug.ResetModule();
-                    Memory.module = Memory.MODULE_OVERTURE_DEBUG;
-                    Memory.IsMouseVisible = false;
-                    init_debugger_Audio.PlayStopMusic();
-                    break;
-
-                case Ditems.Field:
-                    Fade = 0.0f;
-                    Module_field_debug.ResetField();
-                    Memory.module = Memory.MODULE_FIELD_DEBUG;
-                    Memory.IsMouseVisible = false;
-                    break;
-
-                case Ditems.Music:
-                    Module_field_debug.ResetField();
-                    init_debugger_Audio.PlayStopMusic();
-                    break;
-
-                case Ditems.Battle:
-                    Fade = 0.0f;
-                    Memory.battle_encounter = debug_choosedBS;
-                    Module_battle_debug.ResetState();
-                    Memory.module = Memory.MODULE_BATTLE_DEBUG;
-                    Memory.IsMouseVisible = false;
-                    break;
-
-                case Ditems.Sounds:
-                    init_debugger_Audio.PlaySound(debug_choosedAudio);
-                    break;
-
-                case Ditems.Movie:
-                    Fade = 0.0f;
-                    MoviePointer = MoviePointer; //makes movieindex in player match the moviepointer, it is set when ever this is.
-                    Memory.module = Memory.MODULE_MOVIETEST;
-                    Module_movie_test.MovieState = 0;
-                    Memory.IsMouseVisible = false;
-                    break;
-
-                case Ditems.World:
-                    Fade = 0.0f;
-                    Memory.module = Memory.MODULE_WORLD_DEBUG;
-                    Memory.IsMouseVisible = false;
-                    break;
-
-                case Ditems.Faces:
-                    Fade = 0.0f;
-                    Memory.module = Memory.MODULE_FACE_TEST;
-                    break;
-
-                case Ditems.Icons:
-                    Fade = 0.0f;
-                    Memory.module = Memory.MODULE_ICON_TEST;
-                    break;
-
-                case Ditems.Cards:
-                    Fade = 0.0f;
-                    Memory.module = Memory.MODULE_CARD_TEST;
-                    break;
-
-                default:
-                    ret = false;
-                    break;
-            }
-            if (ret && Ditems.Sounds != Dchoose)
-            {
-                init_debugger_Audio.PlaySound(0);
-            }
-            return ret;
-        }
-
-        private static bool UpdateDebugLobbyRIGHT()
-        {
-            bool ret = true;
-            Input.ResetInputLimit();
-            switch (Dchoose)
-            {
-                case Ditems.Sounds:
-                    if (debug_choosedAudio < init_debugger_Audio.soundEntriesCount)
-                    {
-                        debug_choosedAudio++;
-                    }
-
-                    break;
-
-                case Ditems.Battle:
-                    if (debug_choosedBS < Memory.encounters.Length)
-                    {
-                        debug_choosedBS++;
-                    }
-
-                    break;
-
-                case Ditems.Field:
-                    FieldPointer++;
-                    break;
-
-                case Ditems.Movie:
-                    MoviePointer++;
-                    break;
-
-                case Ditems.Music:
-                    //case Ditems.MusicNext:
-                    if (Memory.MusicIndex >= ushort.MaxValue || Memory.MusicIndex >= Memory.dicMusic.Keys.Max())
-                    {
-                        Memory.MusicIndex = 0;
-                    }
-                    else
-                    {
-                        Memory.MusicIndex++;
-                    }
-                    debug_choosedMusic = Path.GetFileNameWithoutExtension(Memory.dicMusic[Memory.MusicIndex][0]);
-                    //init_debugger_Audio.PlayMusic();
-                    break;
-
-                default:
-                    ret = false;
-                    break;
-            }
-            if (ret)
-            {
-                init_debugger_Audio.PlaySound(0);
-            }
-
-            return ret;
-        }
-
-        /// <summary>
-        /// Trigger required draw function.
-        /// </summary>
-        internal static void Draw()
-        {
-            Memory.graphics.GraphicsDevice.Clear(Color.Black);
-            //fScaleWidth = (float)Memory.graphics.GraphicsDevice.Viewport.Width / Memory.PreferredViewportWidth;
-            //fScaleHeight = (float)Memory.graphics.GraphicsDevice.Viewport.Height / Memory.PreferredViewportHeight;
-            vpWidth = Memory.PreferredViewportWidth;//Memory.graphics.GraphicsDevice.Viewport.Width;
-            vpHeight = Memory.PreferredViewportHeight;//Memory.graphics.GraphicsDevice.Viewport.Width;
-            lastfade = fade;
-            vpSpace = vpHeight * 0.09f * Memory.Scale().X;
-            DFontPos = new Vector2(vpWidth * .10f * Memory.Scale().X, vpHeight * .05f * Memory.Scale().Y) + Offset;
-            switch (State)
-            {
-                case MainMenuStates.Init:
-                case MainMenuStates.MainLobby:
-                    DrawMainLobby();
-                    break;
-
-                case MainMenuStates.DebugScreen:
-                    DrawDebugLobby();
-                    break;
-
-                case MainMenuStates.NewGameChoosed:
-                    DrawMainLobby();
-                    break;
-
-                case MainMenuStates.LoadGameLoading:
-                    break;
-
-                case MainMenuStates.LoadGameScreen:
-                    break;
-            }
-        }
-
-        /// <summary>
-        /// Draw Main menu
-        /// </summary>
-        private static void DrawMainLobby()
-        {
-            Vector2 scale = Memory.Scale();
-            Vector2 vp = (new Vector2(vpWidth, vpHeight)) * scale;
-            float item = 0;
-            Vector2 textSize = new Vector2(2f, 3f);// scaled in render function.
-            Vector2 textStart = new Vector2(vp.X * .41f, vp.Y * .65f);
-            Memory.SpriteBatchStartAlpha();
-            Rectangle dst = new Rectangle()
-            {
-                Location = new Point(0),
-                Size = vp.ToPoint()
-            };
-            start.Draw(dst, null, Color.White * fade);
-            foreach (Mitems i in (Mitems[])Enum.GetValues(typeof(Mitems)))
-            {
-                Item c = strMainLobby[i];
-                c.Loc = (Memory.font.RenderBasicText(c.Text.Replace("\0", ""),
-                    (int)(textStart.X), (int)(textStart.Y + ((textSize.Y + vpSpace) * item++)), textSize.X, textSize.Y, 0, 1, Fade));
-                strMainLobby[i] = c;
-            }
-            Memory.SpriteBatchEnd();
-            Memory.SpriteBatchStartAlpha(SamplerState.PointClamp);
-            Memory.Icons.Draw(Icons.ID.Finger_Right, 2, new Rectangle(
-                (int)(textStart.X - ((48 + 10) * scale.X)),
-                (int)(textStart.Y + (10 * scale.Y) + ((textSize.Y + vpSpace) * (float)Mchoose)),
-                (int)(48 * scale.X),
-                (int)(36 * scale.Y)), 0, fade);
-            Memory.SpriteBatchEnd();
         }
 
         private static Rectangle FontBoxCalc<T>(Dictionary<Enum, Item> dict)
@@ -726,7 +209,11 @@ namespace FF8
             foreach (Enum i in Enum.GetValues(typeof(T)))
             {
                 Item c = dict[i];
-                c.Loc = Memory.font.CalcBasicTextArea(string.Format(c.Text, Font.CipherDirty(InfoForLobby<T>(i)).Replace("\0", "")),
+                byte[] end = Font.CipherDirty(InfoForLobby<Ditems>(i));
+                byte[] combine = new byte[strDebugLobby[i].Text.Length + end.Length];
+                Array.Copy(strDebugLobby[i].Text, combine, strDebugLobby[i].Text.Length);
+                Array.Copy(end, 0, combine, strDebugLobby[i].Text.Length, end.Length);
+                c.Loc = Memory.font.CalcBasicTextArea(combine,
                 (int)DFontPos.X, (int)(DFontPos.Y + vpSpace * item++), 1f, 2f, 0);
                 if (dst.X == 0 || dst.Y == 0)
                     dst.Location = c.Loc.Location;
@@ -741,87 +228,16 @@ namespace FF8
         }
 
         /// <summary>
-        /// Draw Debug Menu
+        /// Init
         /// </summary>
-        private static void DrawDebugLobby()
+        private static void Init()
         {
-            Vector2 scale = Memory.Scale();
-            vpSpace = vpHeight * 0.05f * scale.X;
-            float item = 0;
-            Rectangle dst = FontBoxCalc<Ditems>(strDebugLobby);
-            Memory.SpriteBatchStartAlpha(SamplerState.PointClamp);
-            Memory.Icons.Draw(Icons.ID.Menu_BG_256, 0, dst, 0, Fade);
-            item = 0;
-            dst.Offset(4 * 3.5f, 0);
-            dst.Size = (Memory.Icons[Icons.ID.DEBUG].GetRectangle.Size.ToVector2() * scale * 3.5f).ToPoint();
-            Memory.Icons.Draw(Icons.ID.DEBUG, 2, dst, 0, fade);
-            dst.Location = DFontPos.ToPoint();
-            dst.Size = new Point((int)(24 * 2 * scale.X), (int)(16 * 2 * scale.Y));
-            dst.Offset(-(dst.Width + 10 * scale.X), 6 * scale.Y + vpSpace * ((float)Dchoose));
-
-            Memory.Icons.Draw(Icons.ID.Finger_Right, 2, dst, 0, fade);
-            Memory.SpriteBatchEnd();
-            //pointclamp looks bad on default fonts.
-            Memory.SpriteBatchStartAlpha();
-            foreach (Ditems i in (Ditems[])Enum.GetValues(typeof(Ditems)))
-            {
-                Memory.font.RenderBasicText(string.Format(strDebugLobby[i].Text, Font.CipherDirty(InfoForLobby<Ditems>(i)).Replace("\0", "")),
-                    (int)(DFontPos.X), (int)(DFontPos.Y + vpSpace * item++), 1f, 2f, 0, 1, Fade);
-            }
-            //Memory.spriteBatch.Draw(Memory.iconsTex[2], dst,
-            //    new Rectangle(232, 0, 23, 15), Color.White * Fade);
-            Memory.SpriteBatchEnd();
-        }
-
-        /// <summary>
-        /// Dynamic info for Ditem that is read at draw time.
-        /// </summary>
-        /// <param name="i">Ditem being drawn</param>
-        /// <returns>Dynamic info for Ditem</returns>
-        private static string InfoForLobby<T>(Enum i)
-        {
-            switch (typeof(T).Name)
-            {
-                case "Ditems":
-                    switch (i)
-                    {
-                        case Ditems.Battle:
-                            return debug_choosedBS.ToString("D4");
-
-                        case Ditems.Field:
-                            return debug_choosedField;
-
-                        case Ditems.Movie:
-                            return debug_choosedMovie;
-
-                        case Ditems.Sounds:
-                            return $"{debug_choosedAudio}";
-
-                        case Ditems.Music:
-                            return $"{debug_choosedMusic}";
-                    };
-                    break;
-            };
-            return "";
+            InitMain();
+            InitLoad();
+            InitDebug();
+            Memory.Strings.Close();
         }
 
         #endregion Methods
-
-        #region Structs
-
-        /// <summary>
-        /// Container for MenuItems containing relevant info.
-        /// </summary>
-        private struct Item
-        {
-            #region Properties
-
-            public Rectangle Loc { get; set; }
-            public string Text { get; set; }
-
-            #endregion Properties
-        }
-
-        #endregion Structs
     }
 }
